Indexer and Tokenizer now in different files
